+++ backend/src/main/java/com/foureyes/moai/backend/domain/schedule/dto/response/GetScheduleListDto.java
@@ -14,13 +14,13 @@ import java.time.LocalDateTime;
 public class GetScheduleListDto {
 
     @Schema(description = "일정 ID", example = "1")
-    private Long id;
+    private int id;
 
     @Schema(description = "스터디 ID", example = "5")
-    private Long studyId;
+    private int studyId;
 
     @Schema(description = "작성자 사용자 ID", example = "3")
-    private Long userId;
+    private int userId;
 
     @Schema(description = "시작 시간", example = "2025-12-08T09:00:00")
     private LocalDateTime startDatetime;
@@ -34,8 +34,8 @@ public class GetScheduleListDto {
     public static GetScheduleListDto from(Schedule s) {
         return GetScheduleListDto.builder()
             .id(s.getId())
-            .studyId((long) s.getStudyGroup().getId())
-            .userId((long) s.getUser().getId())
+            .studyId(s.getStudyGroup().getId())
+            .userId(s.getUser().getId())
             .startDatetime(s.getStartDatetime())
             .endDatetime(s.getEndDatetime())
             .title(s.getTitle())
