+++ backend/src/main/java/com/foureyes/moai/backend/domain/schedule/repository/ScheduleRepository.java
@@ -4,13 +4,14 @@ import com.foureyes.moai.backend.domain.schedule.entity.Schedule;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
+import software.amazon.awssdk.services.s3.endpoints.internal.Value;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
 public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
-    Optional<Schedule> findById(Long id);
+    Optional<Schedule> findById(Integer id);
 
     @Query("""
       select s from Schedule s
@@ -19,7 +20,7 @@ public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
         and s.endDatetime   > :from
       order by s.startDatetime asc
     """)
-    List<Schedule> findOverlapping(@Param("studyId") Long studyId,
+    List<Schedule> findOverlapping(@Param("studyId") Integer studyId,
                                    @Param("from") LocalDateTime from,
                                    @Param("to") LocalDateTime to);
 }
